# backend/openapi.yaml
openapi: 3.0.3
info:
  title: AIPPOF API - 个人养老金优化计算平台
  description: |
    AIPPOF (AI-Powered Individual Pension Optimization Framework) API文档
    
    本API提供两条主要路径:
    - **PathA (新参与者)**: 为首次参与者提供缴费方案优化
    - **PathB (已参与者)**: 为已有缴费历史的参与者提供诊断和优化建议
    
    核心功能:
    - T2税收优惠计算 (兰浩哥公式)
    - T3领取期税率计算 (双逻辑函数)
    - 精准补贴计算 (两档匹配+削减机制)
    - 缴费上限计算 (Formula 5-5混合动态上限)
    - AI工资增长预测
    - NPV收益计算
    - 历史缴费诊断
    - 个性化优化建议
    
  version: 1.0.0
  contact:
    name: AIPPOF Development Team
    email: support@aippof.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: 本地开发服务器
  - url: https://api.aippof.com
    description: 生产环境服务器
...
paths:
...
  /api/predict-wage-growth:
    post:
      tags:
        - PathA - 新参与者
      summary: 工资增长率预测
      description: |
        使用AI和联网搜索预测未来工资增长率
        
        预测方法:
        - 基础预测 (行业+岗位+年龄+收入因子)
        - AI深度思考 (可选,需要OPENAI_API_KEY)
        - 联网搜索验证 (可选)
        
        最终预测 = 基础40% + AI40% + 搜索20%
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - age
                - annualSalary
                - industry
                - jobLevel
              properties:
                age:
                  type: integer
                  minimum: 18
                  maximum: 65
                  description: 当前年龄
                  example: 30
                annualSalary:
                  type: number
                  minimum: 0
                  description: 年薪 (元)
                  example: 150000
                industry:
                  type: string
                  enum: [it, finance, manufacturing, education, healthcare, other]
                  description: 所属行业
                  example: "it"
                jobLevel:
                  type: string
                  enum: [entry, intermediate, senior, expert, management]
                  description: 职级
                  example: "intermediate"
      responses:
        '200':
          description: 预测成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  predicted_growth_rate:
                    type: number
                    description: 预测增长率 (%)
                    example: 5.65
                  confidence:
                    type: number
                    description: 置信度 (0-1)
                    example: 0.85
                  industry_average:
                    type: number
                    description: 行业平均增长率 (%)
                    example: 5.0
                  methodology:
                    type: string
                    description: 预测方法
                    example: "base+ai+web"
                  factors:
                    type: object
                    properties:
                      base_growth:
                        type: number
                        example: 5.2
                      ai_adjusted_growth:
                        type: number
                        nullable: true
                        example: 6.1
                      web_search_growth:
                        type: number
                        nullable: true
                        example: 5.5
                  ai_insights:
                    type: array
                    items:
                      type: string
                    example: ["IT行业持续高增长", "中级岗位晋升空间大"]
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/calculate-t2:
    post:
      tags:
        - 核心计算
      summary: T2税收优惠率计算
      description: |
        使用兰浩哥公式计算T2税收优惠率
        
        公式: T2 = (实际税收节约) / 12000 * 100%
        
        考虑因素:
        - 缴费年限 (60-年龄)
        - 工资增长率
        - 边际税率
        - 投资回报率 (1.75%)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - age
                - annualSalary
                - wageGrowthRate
              properties:
                age:
                  type: integer
                  minimum: 18
                  maximum: 59
                  example: 30
                annualSalary:
                  type: number
                  minimum: 0
                  example: 150000
                wageGrowthRate:
                  type: number
                  minimum: 0
                  maximum: 20
                  description: 工资增长率 (%)
                  example: 4.5
      responses:
        '200':
          description: 计算成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  t2:
                    type: number
                    description: T2税收优惠率 (%)
                    example: 10.0
                  t1:
                    type: number
                    description: 边际税率 (%)
                    example: 20.0
                  n:
                    type: integer
                    description: 缴费年限
                    example: 30
                  formula:
                    type: string
                    description: 计算公式
                    example: "T2 = 12000.00 / 12000 * 100% = 10.00%"
                  details:
                    type: object
                    properties:
                      contributionYears:
                        type: integer
                        example: 30
                      returnRate:
                        type: number
                        example: 1.75
                      wageGrowthRate:
                        type: number
                        example: 4.5
                      marginalTaxRate:
                        type: number
                        example: 20.0
                      taxSaving:
                        type: number
                        example: 12000.0
  /api/calculate-t3:
    post:
      tags:
        - 核心计算
      summary: T3领取期税率计算
      description: |
        使用双逻辑函数计算T3领取期税率
        
        公式: T3 = f1(T2) + f2(收入)
        - f1: T2敏感逻辑函数 (0-7%)
        - f2: 收入敏感逻辑函数 (0-7%)
        
        T3范围: 0-14%
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - t2
                - annualSalary
                - age
              properties:
                t2:
                  type: number
                  minimum: 0
                  maximum: 45
                  description: T2税收优惠率 (%)
                  example: 10.0
                annualSalary:
                  type: number
                  minimum: 0
                  example: 150000
                age:
                  type: integer
                  minimum: 18
                  maximum: 65
                  example: 30
      responses:
        '200':
          description: 计算成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  t3:
                    type: number
                    description: T3领取期税率 (%)
                    example: 7.2
                  formula:
                    type: string
                    example: "T3 = f1(T2) + f2(Income) = 7.20%"
                  components:
                    type: object
                    properties:
                      t2_component:
                        type: number
                        example: 3.8
                      income_component:
                        type: number
                        example: 3.4
  /api/calculate-subsidy:
    post:
      tags:
        - 核心计算
      summary: 精准补贴计算
      description: |
        计算政府精准补贴金额
        
        机制:
        - 两档匹配: 低收入50%, 中收入30%
        - 收入削减: 线性削减至归零点
        
        **重要**: 年薪≥15万, 补贴=0
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - annualSalary
                - contributionAmount
              properties:
                annualSalary:
                  type: number
                  minimum: 0
                  example: 60000
                contributionAmount:
                  type: number
                  minimum: 0
                  maximum: 12000
                  example: 5000
      responses:
        '200':
          description: 计算成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  subsidy:
                    type: number
                    description: 补贴金额 (元)
                    example: 612.0
                  tier:
                    type: string
                    enum: [low_income, middle_income, high_income]
                    example: "low_income"
                  match_rate:
                    type: number
                    description: 匹配率 (%)
                    example: 50.0
                  taper_factor:
                    type: number
                    description: 削减系数 (0-1)
                    example: 0.245
  /api/calculate-cap:
    post:
      tags:
        - 核心计算
      summary: 缴费上限计算
      description: |
        使用Formula 5-5计算缴费上限
        
        策略: min(动态上限, 固定上限12000)
        
        动态上限 = f(年薪, T2率)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - annualSalary
                - t2Rate
              properties:
                annualSalary:
                  type: number
                  minimum: 0
                  example: 150000
                t2Rate:
                  type: number
                  minimum: 0
                  maximum: 45
                  example: 10.0
      responses:
        '200':
          description: 计算成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  cap:
                    type: number
                    description: 缴费上限 (元)
                    example: 12000
                  strategy:
                    type: string
                    description: 使用的策略
                    example: "min_dynamic_vs_fixed"
                  details:
                    type: object
                    properties:
                      dynamic_cap:
                        type: number
                        example: 15000
                      fixed_cap:
                        type: number
                        example: 12000
                      selected_cap:
                        type: number
                        example: 12000
  /api/optimize-contribution:
    post:
      tags:
        - PathA - 新参与者
      summary: 缴费方案优化
      description: |
        为新参与者生成3档缴费方案
        
        方案档位:
        - 保守型 (60%上限)
        - 均衡型 (80%上限) ⭐ 推荐
        - 激进型 (100%上限)
        
        每个方案包含: 缴费额、T2、补贴、NPV
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - age
                - annualSalary
                - wageGrowthRate
              properties:
                age:
                  type: integer
                  example: 30
                annualSalary:
                  type: number
                  example: 150000
                wageGrowthRate:
                  type: number
                  example: 4.5
      responses:
        '200':
          description: 优化成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  scenarios:
                    type: array
                    minItems: 3
                    maxItems: 3
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                          example: "均衡型"
                        contribution:
                          type: number
                          example: 12000
                        predictedT2:
                          type: number
                          example: 10.0
                        subsidy:
                          type: number
                          example: 0
                        npv:
                          type: number
                          example: 61961.12
                  t2:
                    type: number
                    example: 10.0
                  t3:
                    type: number
                    example: 7.2
                  cap:
                    type: number
                    example: 12000
                  subsidyTierInfo:
                    type: object

  # ==================== PathB API ====================
...
components:
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: 错误消息
          example: "缺少必填字段: age"
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API密钥认证 (未来版本)
security: []